@@ -1,6 +1,6 @@
 {
   "name": "react-sparklines",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "description": "Beautiful and expressive Sparklines component for React ",
   "main": "build/index.js",
   "directories": {
@@ -36,13 +36,13 @@
   "devDependencies": {
     "babel": "^6.5.2",
     "babel-core": "^6.8.0",
-    "babel-loader": "^6.2.4",
+    "babel-loader": "7.1.1",
     "babel-plugin-transform-object-assign": "^6.8.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-1": "^6.5.0",
     "babel-runtime": "^6.6.1",
-    "chai": "^3.5.0",
+    "chai": "^4.1.0",
     "enzyme": "^2.2.0",
     "hiff": "^0.3.0",
     "line-by-line": "^0.1.4",
@@ -50,9 +50,9 @@
     "react": "^15.0.2",
     "react-addons-test-utils": "^15.0.2",
     "react-dom": "^15.0.2",
-    "react-element-to-jsx-string": "=2.6.1",
+    "react-element-to-jsx-string": "11.0.1",
     "replaceall": "^0.1.6",
-    "webpack": "^2.1.0-beta.22",
+    "webpack": "^3.4.1",
     "webpack-dev-server": "^2.2.0"
   },
   "peerDependencies": {
@@ -60,6 +60,6 @@
     "react-dom": "*"
   },
   "dependencies": {
-    "react-addons-shallow-compare": "^15.0.2"
+    "prop-types": "^15.5.10"
   }
 }
