@@ -1,4 +1,5 @@
-import React from 'react';
+import PropTypes from 'prop-types';
+import React, { PureComponent} from 'react';
 import SparklinesText from './SparklinesText';
 import SparklinesLine from './SparklinesLine';
 import SparklinesCurve from './SparklinesCurve';
@@ -7,22 +8,22 @@ import SparklinesSpots from './SparklinesSpots';
 import SparklinesReferenceLine from './SparklinesReferenceLine';
 import SparklinesNormalBand from './SparklinesNormalBand';
 import dataToPoints from './dataProcessing/dataToPoints';
-import shallowCompare from 'react-addons-shallow-compare';
 
-class Sparklines extends React.Component {
+class Sparklines extends PureComponent {
 
     static propTypes = {
-        data: React.PropTypes.array,
-        limit: React.PropTypes.number,
-        width: React.PropTypes.number,
-        height: React.PropTypes.number,
-        svgWidth: React.PropTypes.number,
-        svgHeight: React.PropTypes.number,
-        preserveAspectRatio: React.PropTypes.string,
-        style: React.PropTypes.object,
-        min: React.PropTypes.number,
-        max: React.PropTypes.number,
-        onMouseMove: React.PropTypes.func
+        data: PropTypes.array,
+        limit: PropTypes.number,
+        width: PropTypes.number,
+        height: PropTypes.number,
+        svgWidth: PropTypes.number,
+        svgHeight: PropTypes.number,
+        preserveAspectRatio: PropTypes.string,
+        margin: PropTypes.number,
+        style: PropTypes.object,
+        min: PropTypes.number,
+        max: PropTypes.number,
+        onMouseMove: PropTypes.func
     };
 
     static defaultProps = {
@@ -35,10 +36,6 @@ class Sparklines extends React.Component {
 
     constructor (props) {
         super(props);
-    }
-
-    shouldComponentUpdate(nextProps) {
-        return shallowCompare(this, nextProps);
     }
 
     render() {
