@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import mean from './dataProcessing/mean';
 import stdev from './dataProcessing/stdev';
@@ -5,7 +6,7 @@ import stdev from './dataProcessing/stdev';
 export default class SparklinesNormalBand extends React.Component {
 
     static propTypes = {
-        style: React.PropTypes.object
+        style: PropTypes.object
     };
 
     static defaultProps = {
