@@ -1,14 +1,17 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 export default class SparklinesSpots extends React.Component {
 
     static propTypes = {
-        size: React.PropTypes.number,
-        style: React.PropTypes.object,
-        spotColors: React.PropTypes.object
+        index: PropTypes.number,
+        size: PropTypes.number,
+        style: PropTypes.object,
+        spotColors: PropTypes.object
     };
 
     static defaultProps = {
+        index: null,
         size: 2,
         spotColors: {
             '-1': 'red',
@@ -18,17 +21,18 @@ export default class SparklinesSpots extends React.Component {
     };
 
     lastDirection(points) {
+        const { index } = this.props;
 
         Math.sign = Math.sign || function(x) { return x > 0 ? 1 : -1; }
 
-        return points.length < 2
+        return index < 2
             ? 0
-            : Math.sign(points[points.length - 2].y - points[points.length - 1].y);
+            : Math.sign(points[index - 2].y - points[index - 1].y);
     }
 
     render() {
 
-        const { points, width, height, size, style, spotColors } = this.props;
+        const { index, points, width, height, size, style, spotColors } = this.props;
 
         const startSpot = <circle
                             cx={points[0].x}
@@ -37,8 +41,8 @@ export default class SparklinesSpots extends React.Component {
                             style={style} />
 
         const endSpot = <circle
-                            cx={points[points.length - 1].x}
-                            cy={points[points.length - 1].y}
+                            cx={points[index].x}
+                            cy={points[index].y}
                             r={size}
                             style={style || { fill: spotColors[this.lastDirection(points)] }} />
 
